pyproject.toml
[project]
name = "bank-homework"
version = "0.1.0"
description = ""
authors = [
    {name = "SHADrussed",email = "shad_rus_teaT@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
pytest = "^8.3.4"
python = "^3.13"
pandas = "2.2.3"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '.venv'

[tool.poetry]
package-mode = false

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '.git'

[tool.isort]
line_length = 119